% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS_map.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot map based on FARS data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A list or numeric vector containing the numeric code of
the desired states to plot.}

\item{year}{A numeric value to indicate the year of the data to plot.}
}
\value{
In case of success the function returns a map with the desired data.
}
\description{
This is a complex function that plots the accidents per state in a map.
}
\details{
The function first builds the name of the data file to be read
via \code{\link{make_filename}}, and then reads the data with \code{\link{fars_read}}.
The state code is converted to an integer, and then is compared to the states
present in the data. A mechanism for errors is implemented, the first
detects invalid state numbers. Then the function filters the data by the states,
and evaluates if there are enough data to plot (>0L). After it, the function plots
the map with the desired data.
}
\examples{
\dontrun{fars_map_state(c(01, 10, 25), 2012)}

}
