% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS_read_M.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data from multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector or a list containing the years of the
data desired.}
}
\value{
The function returns a list of data frames containing the loaded data.
In case invalid outputs are used, the returned value is NULL, prior printing a
warning on the console.
}
\description{
This function reads the FARS data of multiple years, and proceses the
number of incidents per month and year.
}
\details{
This function uses the elements on the vector or list \code{years}
to create filenames through \link{make_filename()} and read each file with
\link{fars_read()}. Then selects the month and year columns of the data.
A mechanism to detect invalid years is implemented to display
warnings when invalid values are used as input.
}
\examples{
\dontrun{a <- c(1998, 2001, 2005)
data <- fars_read_years(a)}
\dontrun{ a <- list(1998, 2001, 2005)
data <- fars_read_years(a)}

}
