% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARS_summary.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Number of Accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector or a list containing the years of the
data desired.}
}
\value{
The function returns a dataframe with n columns (years) and 12 rows (months).
The value of each cell corresponds to the number of accidents.
}
\description{
This function summarizes the number of accident per month and year in the fars data.
}
\details{
The function compiles the data asked with the param \code{years} using
\code{\link{fars_read_years}} and binds each element of the resulting list by rows.
Then the function groups the data per month and year and summarizes the number of observations
(accidents), per year. Finally, the function makes a contingency table by month and year.
}
\examples{
\dontrun{fars_summarize_years(2013)}
\dontrun{a <- list(1998, 2001, 2005)}
\dontrun{data <- fars_read_years(a)}


}
