% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MK_Filename.R
\name{make_filename}
\alias{make_filename}
\title{Make Filenames for FARS data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numerical value representing the year of the desired data.}
}
\value{
The function returns an string with the filename of the FARS data of the
corresponding year.
}
\description{
This is a simple function that creates a name for compressed files of the
FARS data.
}
\details{
The function coerces the argument 'year' to an integer value,
posteriorly builds an string for the default name of the compressed data using
sprintf.
Using other imputs will result in coercion to NA's and therefore, filenames containing
NA's.
}
\examples{
\dontrun{make_filename(1994)}
\dontrun{make_filename(2013)}

}
